Float. D'oh

+++ Damselfly.Core/Services/MetaDataService.cs
@@ -341,7 +341,7 @@ public class MetaDataService : IProcessJobFactory
 
             if( imgMetaData.Width != 0 && imgMetaData.Height != 0 )
             {
-                imgMetaData.AspectRatio = imgMetaData.Width / imgMetaData.Height;
+                imgMetaData.AspectRatio = (double)imgMetaData.Width / (double)imgMetaData.Height;
             }
 
             Logging.Log($"Read metadata for {image.FullPath} (ID: {image.ImageId}) including {keywords.Count()} keywords.");
